refactor: change `server/types/index` to `server/types/general`
